backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
websockets
passlib[bcrypt]
python-jose[cryptography]
multipart